♻️ Refactor: 컴포넌트 별로 디렉토리 생성 및  스타일링 파일 분리
