+++ src/components/@common/Footer/Footer.styles.ts
@@ -1,0 +1,15 @@
+import styled from '@emotion/styled';
+
+export const Container = styled.div`
+  display: flex;
+  justify-content: center;
+  margin-top: auto;
+  padding: 3rem 0;
+  font-size: 0.75rem;
+  text-align: center;
+  line-height: 1.5;
+
+  @media (max-width: 768px) {
+    font-size: 10px;
+  }
+`;
