+++ src/components/@common/Footer/Footer.tsx
@@ -1,0 +1,13 @@
+import React, { FunctionComponent } from 'react';
+import * as S from './Footer.styles';
+
+const Footer: FunctionComponent = function () {
+  return (
+    <S.Container>
+      Thank You for Visiting My Blog, Have a Good Day🌼❤️
+      <br />© 2022 Developer DaisyLee, Powered By Gatsby.
+    </S.Container>
+  );
+};
+
+export default Footer;
