+++ src/components/CategoryList/CategoryList.styles.ts
@@ -1,0 +1,14 @@
+import styled from '@emotion/styled';
+
+export const Container = styled.div`
+  display: flex;
+  flex-wrap: wrap;
+  width: 768px;
+  margin: 100px auto 0;
+
+  @media (max-width: 768px) {
+    width: 100%;
+    margin-top: 50px;
+    padding: 0 20px;
+  }
+`;
