+++ src/components/Introduction/Introduction.styles.ts
@@ -1,0 +1,41 @@
+import styled from '@emotion/styled';
+
+export const Background = styled.div`
+  width: 100%;
+  background-image: linear-gradient(60deg, #29323c 0%, #485563 100%);
+  color: #ffffff;
+`;
+
+export const Container = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-start;
+  width: 768px;
+  height: 400px;
+  margin: 0 auto;
+
+  @media (max-width: 768px) {
+    width: 100%;
+    height: 300px;
+    padding: 0 20px;
+  }
+`;
+
+export const Subtitle = styled.div`
+  font-size: 20px;
+  font-weight: 400;
+
+  @media (max-width: 768px) {
+    font-size: 15px;
+  }
+`;
+
+export const Title = styled.div`
+  font-size: 35px;
+  font-weight: 700;
+
+  @media (max-width: 768px) {
+    font-size: 25px;
+  }
+`;
