--- src/components/Main/ProfileImage.tsx
@@ -1,27 +1,0 @@
-import React, { FunctionComponent } from 'react';
-import styled from '@emotion/styled';
-import { GatsbyImage, IGatsbyImageData } from 'gatsby-plugin-image';
-
-type ProfileImageProps = {
-  profileImage: IGatsbyImageData;
-};
-
-const ProfileImageWrapper = styled(GatsbyImage)`
-  width: 120px;
-  height: 120px;
-  margin-bottom: 30px;
-  border-radius: 50%;
-
-  @media (max-width: 768px) {
-    width: 80px;
-    height: 80px;
-  }
-`;
-
-const ProfileImage: FunctionComponent<ProfileImageProps> = function ({
-  profileImage,
-}) {
-  return <ProfileImageWrapper image={profileImage} alt="Profile Image" />;
-};
-
-export default ProfileImage;
