+++ src/components/PostItem/PostItem.styles.ts
@@ -1,13 +1,8 @@
-import React, { FunctionComponent } from 'react';
 import styled from '@emotion/styled';
 import { Link } from 'gatsby';
-import { PostFrontmatterType } from 'types/PostItem.types';
 import { GatsbyImage } from 'gatsby-plugin-image';
 
-//PostItem이 받아오는 props: 제목, 날짜, 카테고리, 요약, 썸네일, 링크
-type PostItemProps = PostFrontmatterType & { link: string };
-
-const PostItemWrapper = styled(Link)`
+export const Container = styled(Link)`
   display: flex;
   flex-direction: column;
   border-radius: 10px;
@@ -22,14 +17,14 @@ const PostItemWrapper = styled(Link)`
 `;
 
 //추가 학습이 필요한 CSS 속성 또는 속성값: object-fit
-const ThumbnailImage = styled(GatsbyImage)`
+export const ThumbnailImage = styled(GatsbyImage)`
   width: 100%;
   height: 200px;
   border-radius: 10px 10px 0 0;
   object-fit: cover;
 `;
 
-const PostItemContent = styled.div`
+export const PostItemContent = styled.div`
   flex: 1;
   display: flex;
   flex-direction: column;
@@ -39,7 +34,7 @@ const PostItemContent = styled.div`
 //추가 학습이 필요한 CSS 속성 또는 속성값: -webkit-box, text-overflow,
 //white-space, overflow-wrap, -webkit-line-clamp,
 //-webkit-box-orient
-const Title = styled.div`
+export const Title = styled.div`
   display: -webkit-box;
   overflow: hidden;
   margin-bottom: 3px;
@@ -52,20 +47,20 @@ const Title = styled.div`
   font-weight: 700;
 `;
 
-const Date = styled.div`
+export const Date = styled.div`
   font-size: 14px;
   font-weight: 400;
   opacity: 0.7;
 `;
 
-const Category = styled.div`
+export const Category = styled.div`
   display: flex;
   flex-wrap: wrap;
   margin-top: 10px;
   margin: 10px -5px;
 `;
 
-const CategoryItem = styled.div`
+export const CategoryItem = styled.div`
   margin: 2.5px 5px;
   padding: 3px 5px;
   border-radius: 3px;
@@ -76,7 +71,7 @@ const CategoryItem = styled.div`
 `;
 
 //추가 학습이 필요한 CSS 속성 또는 속성값: Title 컴포넌트와 동일
-const Summary = styled.div`
+export const Summary = styled.div`
   display: -webkit-box;
   overflow: hidden;
   margin-top: auto;
@@ -88,32 +83,3 @@ const Summary = styled.div`
   font-size: 16px;
   opacity: 0.8;
 `;
-
-const PostItem: FunctionComponent<PostItemProps> = function ({
-  title,
-  date,
-  categories,
-  summary,
-  thumbnail: {
-    childImageSharp: { gatsbyImageData },
-  },
-  link,
-}) {
-  return (
-    <PostItemWrapper to={link}>
-      <ThumbnailImage image={gatsbyImageData} alt="Post Item Image" />
-      <PostItemContent>
-        <Title>{title}</Title>
-        <Date>{date}</Date>
-        <Category>
-          {categories.map(category => (
-            <CategoryItem key={category}>{category}</CategoryItem>
-          ))}
-        </Category>
-        <Summary>{summary}</Summary>
-      </PostItemContent>
-    </PostItemWrapper>
-  );
-};
-
-export default PostItem;
