+++ src/components/PostList/PostList.styles.ts
@@ -1,0 +1,17 @@
+import styled from '@emotion/styled';
+
+export const Container = styled.div`
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  grid-gap: 20px;
+  width: 768px;
+  margin: 0 auto;
+  padding: 50px 0;
+  background-color: tomato;
+
+  @media (max-width: 768px) {
+    grid-template-columns: 1fr;
+    width: 100%;
+    padding: 50px 20px;
+  }
+`;
