+++ src/components/PostList/PostList.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent } from 'react';
-import styled from '@emotion/styled';
-import PostItem from './PostItem';
+import PostItem from '../PostItem/PostItem';
 import { PostListItemType } from 'types/PostItem.types';
+import * as S from './PostList.styles';
 
 export type PostType = {
   node: {
@@ -24,29 +24,13 @@ type PostListProps = {
   posts: PostListItemType[];
 };
 
-const PostListWrapper = styled.div`
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  grid-gap: 20px;
-  width: 768px;
-  margin: 0 auto;
-  padding: 50px 0;
-  background-color: tomato;
-
-  @media (max-width: 768px) {
-    grid-template-columns: 1fr;
-    width: 100%;
-    padding: 50px 20px;
-  }
-`;
-
 const PostList: FunctionComponent<PostListProps> = function ({ posts }) {
   return (
-    <PostListWrapper>
+    <S.Container>
       {posts.map(({ node: { id, frontmatter } }: PostListItemType) => (
         <PostItem {...frontmatter} link="https://www.google.co.kr" key={id} />
       ))}
-    </PostListWrapper>
+    </S.Container>
   );
 };
