+++ src/components/ProfileImage/ProfileImage.styles.ts
@@ -1,0 +1,14 @@
+import styled from '@emotion/styled';
+import { GatsbyImage } from 'gatsby-plugin-image';
+
+export const Container = styled(GatsbyImage)`
+  width: 120px;
+  height: 120px;
+  margin-bottom: 30px;
+  border-radius: 50%;
+
+  @media (max-width: 768px) {
+    width: 80px;
+    height: 80px;
+  }
+`;
