+++ src/components/index.ts
@@ -1,0 +1,5 @@
+export { default as CategoryList } from './CategoryList/CategoryList';
+export { default as Introduction } from './Introduction/Introduction';
+export { default as PostItem } from './PostItem/PostItem';
+export { default as PostList } from './PostList/PostList';
+export { default as ProfileImage } from './ProfileImage/ProfileImage';
